+++ code/threads/list.h
@@ -27,11 +27,11 @@
 class ListElement
 {
   public:
-    ListElement (void *itemPtr, int sortKey);	// initialize a list element
+    ListElement (void *itemPtr, long long sortKey);	// initialize a list element
 
     ListElement *next;		// next element on list, 
     // NULL if this is the last
-    int key;			// priority, for a sorted list
+    long long key;			// priority, for a sorted list
     void *item;			// pointer to item on the list
 };
 
@@ -57,8 +57,8 @@ class List
 
 
     // Routines to put/get items on/off list in order (sorted by key)
-    void SortedInsert (void *item, int sortKey);	// Put item into list
-    void *SortedRemove (int *keyPtr);	// Remove first item from list
+    void SortedInsert (void *item, long long sortKey);	// Put item into list
+    void *SortedRemove (long long *keyPtr);	// Remove first item from list
 
   private:
       ListElement * first;	// Head of the list, NULL if list is empty
